test: add read-only acceptance tests for Admin data sources
Now that the test key targets the dedicated terraform-tests workspace, add
live smoke acceptance tests (gate PreCheckAdmin, target the workspace via
acctest.TerraformTestsWorkspaceID) for the Admin read-only data sources that
lacked them: anthropic_workspace, anthropic_workspaces,
anthropic_workspace_members, anthropic_api_keys.

These complement — they do not replace — the httptest-based unit tests, which
keep deterministic coverage of pagination, query-param/filter construction,
mapping edge cases, and 404 handling that a live API cannot exercise reliably.
Assertions check attribute presence (e.g. members.#) rather than non-
deterministic live values.

Following the repo convention, the external acceptance tests use the plain
<name>_test.go name and the pre-existing internal unit suites are renamed to
the _mapping_test.go suffix (the split is forced: acc tests must live in the
external package to avoid the acctest->provider->service import cycle, while
the unit tests need unexported symbols).
