+++ CLAUDE.md
@@ -83,12 +83,12 @@ internal/
 **Go acceptance tests** (`internal/services/<service>/`):
 - Test files use `package <service>_test` (external test package), except tests that access unexported symbols which use `package <service>`
 - **File naming:** use `<name>_test.go` (e.g. `workspace_rate_limits_data_source_test.go`). Do **not** suffix unit-test files with `_unit_test.go` — acceptance test functions are already disambiguated by the `TestAccXxx` prefix, and unit tests use `TestXxx`.
-- **Important:** acceptance tests (those importing `internal/acctest`) MUST live in `package <service>_test` (external) to avoid an import cycle (`acctest` → `provider` → `<service>`). When unit tests need internal types AND a service has acceptance tests, split them into two files: one in `package <service>` (internal, unit) and one in `package <service>_test` (external, acceptance). When the split is for a **single** resource/data source (so the `<name>` prefix can't disambiguate, unlike `workspace` vs `workspaces`), give the internal-unit file a `_mapping_test.go` suffix and keep the plain `_test.go` for the external acceptance file — e.g. `agent_resource_mapping_test.go` (internal) + `agent_resource_test.go` (acceptance), and `organization_data_source_mapping_test.go` (internal) + `organization_data_source_test.go` (acceptance). Test the inline API-response → state mapping by extracting a `map<Resource>ToState` helper the internal test calls directly.
+- **Important:** acceptance tests (those importing `internal/acctest`) MUST live in `package <service>_test` (external) to avoid an import cycle (`acctest` → `provider` → `<service>`). When unit tests need internal types AND a service has acceptance tests, split them into two files: one in `package <service>` (internal, unit) and one in `package <service>_test` (external, acceptance). When the split is for a **single** resource/data source (so the `<name>` prefix can't disambiguate, unlike `workspace` vs `workspaces`), give the internal-unit file a `_mapping_test.go` suffix and keep the plain `_test.go` for the external acceptance file — e.g. `agent_resource_mapping_test.go` (internal) + `agent_resource_test.go` (acceptance), and `organization_data_source_mapping_test.go` (internal) + `organization_data_source_test.go` (acceptance). Test the inline API-response → state mapping by extracting a `map<Resource>ToState` helper the internal test calls directly. The `_mapping_test.go` suffix marks the internal-unit file generally — it may hold more than mapping (pagination, filter/query-param construction, 404 handling), as in `workspaces_data_source_mapping_test.go` — while the plain `<name>_test.go` always holds the external acceptance tests. Never use an `_acc_test.go` suffix: the `TestAcc` prefix already disambiguates acceptance functions.
 - `internal/acctest/acctest.go` exports `ProtoV6ProviderFactories`, `PreCheck` (standard API key), and `PreCheckAdmin` (admin API key) used by all acceptance tests
 - **Test workspace isolation.** All acceptance/native tests operate within the dedicated `terraform-tests` workspace so test resources never touch production. The standard `ANTHROPIC_API_KEY` used for tests is **scoped to `terraform-tests`**, so standard-API resources created during tests land there automatically. Admin API tests (organization-wide) target it by ID via the single shared constant `acctest.TerraformTestsWorkspaceID` — use that constant, never re-hardcode the `wrkspc_...` literal in Go tests.
 - For an admin `*admin.Client` pointed at an `httptest` server in any package's unit tests, use `admintest.NewClient(t, srv)` (`internal/admintest`) — do not construct `admin.Client` inline or duplicate the helper per package
 - `internal/services/workspaces/workspacetest.go` defines workspaces-specific shared helpers (`workspaceFixture`, `pageData`, `fetchAllPages`, plus a thin `newTestAdminClient` that delegates to `admintest.NewClient`); reuse them in any new workspaces unit test instead of duplicating pagination loops
-- **Admin API acceptance tests for read-only data sources:** target the `terraform-tests` workspace via `acctest.TerraformTestsWorkspaceID` and gate on `acctest.PreCheckAdmin`. Resource tests (create/update/delete) on the Admin API remain blocked by [#58](https://github.com/ippontech/terraform-provider-anthropic/issues/58) because we do not yet have a dedicated test organization, so do not create new resources via Admin API in tests.
+- **Admin API acceptance tests for read-only data sources:** target the `terraform-tests` workspace via `acctest.TerraformTestsWorkspaceID` and gate on `acctest.PreCheckAdmin`. These are **smoke** tests (the live workspace's data isn't deterministic, so assert attribute presence like `members.#`, not specific values); they **complement, not replace**, the httptest-based unit tests that deterministically cover pagination, query-param/filter construction, mapping edge cases, and 404 handling. Covered so far: `workspace`, `workspaces`, `workspace_members`, `workspace_rate_limits`, `organization`, `api_keys`. Resource tests (create/update/delete) on the Admin API remain blocked by [#58](https://github.com/ippontech/terraform-provider-anthropic/issues/58) because we do not yet have a dedicated test organization, so do not create new resources via Admin API in tests.
 - Admin API **Terraform native tests**: use `command = plan` (not the default `apply`) until a test org is available; this validates schema without making live API calls. Read-only data source native tests may run against the `terraform-tests` workspace ID above.
 - Unit tests: no special env vars needed
 - Acceptance tests: use `resource.Test(t, resource.TestCase{...})` with `TF_ACC=1`
